+++ AutoTrader/TechUtils.cpp
@@ -15,7 +15,7 @@ TechUtils::~TechUtils()
 }
 
 
-double TechUtils::CalulateMA(const std::list<TickWrapper>& data, const TickWrapper& current, int seconds)
+double TechUtils::CalulateMA(const std::list<TickWrapper>& data, const TickWrapper& current, size_t seconds)
 {
 	//datetime to timestamp
 	double totalExchangeLastPrice = current.LastPrice();
@@ -39,13 +39,13 @@ double TechUtils::CalulateMA(const std::list<TickWrapper>& data, const TickWrapp
 	return totalExchangeLastPrice / count;
 }
 
-double TechUtils::CalulateEMA(const std::list<TickWrapper>& data, const TickWrapper& current, int seconds)
+double TechUtils::CalulateEMA(const std::list<TickWrapper>& data, const TickWrapper& current, size_t seconds)
 {
 	if (data.empty()){
 		return current.LastPrice();
 	}
 
-	int N = seconds * 2;
+	size_t N = seconds * 2;
 
 	TickWrapper preNode = data.front();
 	MACrossTech* preTechVec = dynamic_cast<MACrossTech*>(preNode.m_techvec);
@@ -68,10 +68,10 @@ double TechUtils::CalulateEMA(const std::list<TickWrapper>& data, const TickWrap
 	}
 }
 
-double TechUtils::CalulateWMA(const std::list<TickWrapper>& data, const TickWrapper& current, int seconds)
+double TechUtils::CalulateWMA(const std::list<TickWrapper>& data, const TickWrapper& current, size_t seconds)
 {
 	//datetime to timestamp
-	int n = seconds * 2;
+	size_t n = seconds * 2;
 	double totalExchangeLastPrice = current.LastPrice() * n;
 	long long count = n--;
 
@@ -95,7 +95,7 @@ double TechUtils::CalulateWMA(const std::list<TickWrapper>& data, const TickWrap
 	return totalExchangeLastPrice / count;
 }
 
-double TechUtils::CalulateAMA(const std::list<TickWrapper>& data, const TickWrapper& current, int seconds)
+double TechUtils::CalulateAMA(const std::list<TickWrapper>& data, const TickWrapper& current, size_t seconds)
 {
 	double totalExchangePrice = current.TurnOver();
 	long long totalVolume = current.Volume();
@@ -119,7 +119,7 @@ double TechUtils::CalulateAMA(const std::list<TickWrapper>& data, const TickWrap
 }
 
 
-double TechUtils::CalulateMA(const std::vector<KData>& data, const KData& current, int mins){
+double TechUtils::CalulateMA(const std::vector<KData>& data, const KData& current, size_t mins){
 	//datetime to timestamp
 	double totalExchangeLastPrice = current.LastPrice();
 	long long count = 1;
@@ -142,7 +142,7 @@ double TechUtils::CalulateMA(const std::vector<KData>& data, const KData& curren
 	return totalExchangeLastPrice / count;
 }
 
-double TechUtils::CalulateEMA(const std::vector<KData>& data, const KData& current, int mins){
+double TechUtils::CalulateEMA(const std::vector<KData>& data, const KData& current, size_t mins){
 	return 0.0;
 	//if (data.empty()){
 	//	return current.LastPrice();
@@ -169,7 +169,7 @@ double TechUtils::CalulateEMA(const std::vector<KData>& data, const KData& curre
 	//}
 }
 
-double TechUtils::CalulateWMA(const std::vector<KData>& data, const KData& current, int mins){
+double TechUtils::CalulateWMA(const std::vector<KData>& data, const KData& current, size_t mins){
 
 	double totalExchangeLastPrice = current.LastPrice() * mins;
 	long long count = mins--;
@@ -194,7 +194,7 @@ double TechUtils::CalulateWMA(const std::vector<KData>& data, const KData& curre
 	return totalExchangeLastPrice / count;
 }
 
-double TechUtils::CalulateAMA(const std::vector<KData>& data, const KData& current, int mins){
+double TechUtils::CalulateAMA(const std::vector<KData>& data, const KData& current, size_t mins){
 	double totalExchangePrice = current.TurnOver();
 	long long totalVolume = current.Volume();
